Move IPAddr refinement to separate file
