+++ lib/pharos/phases/gather_facts.rb
@@ -5,6 +5,8 @@ require 'ipaddr'
 module Pharos
   module Phases
     class GatherFacts < Pharos::Phase
+      using Pharos::CoreExt::IPAddrLoopback if RUBY_VERSION < '2.5.0'
+
       title "Gather host facts"
 
       def call
@@ -101,25 +103,6 @@ module Pharos
       def resolvconf_nameservers
         @ssh.file('/etc/resolv.conf').lines.map { |l| l[/^nameserver ([\h:.]+)/, 1] }.compact
       end
-
-      module IPAddrLoopbackBackport
-        refine IPAddr do
-          # Backported from Ruby 2.5
-          # Returns true if the ipaddr is a loopback address.
-          def loopback?
-            case @family
-            when Socket::AF_INET
-              @addr & 0xff000000 == 0x7f000000
-            when Socket::AF_INET6
-              @addr == 1
-            else
-              false
-            end
-          end
-        end
-      end
-
-      using IPAddrLoopbackBackport if RUBY_VERSION < '2.5.0'
 
       # @return [Boolean]
       def resolvconf_nameserver_localhost?
